Implemented styling refactoring for various components #56
