+++ __tests__/campus-toggle.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import CampusToggle from "../src/components/campus-toggle/campus-toggle.component";
-import { CONCORDIA_RED, INACTIVE_BUTTON_COLOR } from "../src/constants/style";
+import { CONCORDIA_RED, INACTIVE_BUTTON_COLOR } from "../src/styles/colors";
 import renderer from "react-test-renderer";
 import { shallow } from "enzyme";
 import { getCampusById } from "../src/constants/campus.data";
