+++ src/styles/button.styles.ts
@@ -1,0 +1,9 @@
+export const Circle: Object = {
+  borderWidth: 1,
+  borderColor: "rgba(0,0,0,0.2)",
+  alignItems: "center",
+  justifyContent: "center",
+  width: 60,
+  height: 60,
+  borderRadius: 50
+};
