+++ src/styles/colors.styles.ts
@@ -1,0 +1,15 @@
+export const BUILDING_MARKER_COLOR = "#252525";
+export const BUILDING_UNTAPPED = "#dba4a8";
+export const BUTTON_BORDER_COLOR = "rgba(0,0,0,0.2)";
+export const CONCORDIA_RED = "#AA2B45";
+export const INACTIVE_BUTTON_COLOR = "#F2F2F2";
+export const INPUT_BORDER_COLOR = "#DEDEDE";
+export const LIST_BACKGROUND_COLOR = "#F7F7FA";
+export const LIST_ITEM_BORDER_COLOR = "rgba(120, 132, 158, 0.08)";
+export const LIST_ITEM_TEXT_COLOR = "#454F63";
+export const WHITE_BACKGROUND_COLOR = "white";
+export const INACTIVE_TEXT_COLOR = "#717171";
+export const INACTIVE_ICON_COLOR = "#C76969";
+export const CUSTOM_MARKER_COLOR = "#f0f0f0";
+export const LOCATION_BUTTON_COLOR = "#fff";
+export const DEFAULT_STACK_NAVIGATOR_COLOR = "#9f1d35";
