+++ src/styles/position-styling.styles.ts
@@ -1,0 +1,6 @@
+export const CAMPUS_TOGGLE_HEIGHT = 80;
+export const MAP_BUTTON_RIGHT = 10;
+export const BUILDING_LOCATION_HEIGHT = CAMPUS_TOGGLE_HEIGHT + 30;
+export const START_TRAVEL_HEIGHT = 40;
+export const FLOOR_PICKER_HEIGHT = BUILDING_LOCATION_HEIGHT + 80;
+export const FLOOR_PICKER_TRAVEL_HEIGHT = START_TRAVEL_HEIGHT + 80;
