+++ src/styles/text.styles.ts
@@ -1,0 +1,1 @@
+export const CURRENT_LOCATION_DISPLAY_TEXT = "Current Location";
